Get rid of ChatHelper
