+++ extensions/wikia/Chat2/ChatController.class.php
@@ -9,7 +9,7 @@ class ChatController extends WikiaController {
 	public function executeIndex() {
 		global $wgUser, $wgFavicon, $wgOut, $wgHooks;
 
-		ChatHelper::info( __METHOD__ . ': Method called' );
+		Chat::info( __METHOD__ . ': Method called' );
 		wfProfileIn( __METHOD__ );
 
 		// String replacement logic taken from includes/Skin.php
