+++ extensions/wikia/Chat2/ChatServerApiClient.class.php
@@ -35,14 +35,14 @@ class ChatServerApiClient {
 			$cityId = $cityData->cityId;
 		} else {
 			// FIXME: How should we handle it if there is no cityId?
-			ChatHelper::info( __METHOD__ . ': Method called - no cityId', [
+			Chat::info( __METHOD__ . ': Method called - no cityId', [
 				'roomId' => $roomId,
 			] );
 		}
 
 		wfProfileOut( __METHOD__ );
 
-		ChatHelper::info( __METHOD__ . ': Method called', [
+		Chat::info( __METHOD__ . ': Method called', [
 			'roomId' => $roomId,
 			'cityId' => $cityId,
 		] );
@@ -101,12 +101,12 @@ class ChatServerApiClient {
 
 		if ( isset( $roomData->roomId ) ) {
 			$roomId = $roomData->roomId;
-			ChatHelper::info( __METHOD__ . ': Method called', [
+			Chat::info( __METHOD__ . ': Method called', [
 				'roomId' => $roomId,
 			] );
 		} else {
 			// FIXME: How should we handle it if there is no roomId?
-			ChatHelper::info( __METHOD__ . ': Method called - no roomId' );
+			Chat::info( __METHOD__ . ': Method called - no roomId' );
 		}
 
 		wfProfileOut( __METHOD__ );
@@ -126,7 +126,7 @@ class ChatServerApiClient {
 
 		$response = null;
 
-		ChatHelper::debug( __METHOD__ . ': Method called ', $params );
+		Chat::debug( __METHOD__ . ': Method called ', $params );
 
 		// NOTE: When we fail over, the chat server host doesn't change which backend it points to (since there isn't
 		// even a chat server in the backup datacenter(s)), so if we're in read-only, even though this isn't a write
