+++ extensions/wikia/Chat2/SpecialChat.class.php
@@ -13,19 +13,19 @@ class SpecialChat extends UnlistedSpecialPage {
 		// check if logged in
 		if ( $wgUser->isLoggedIn() ) {
 			if ( Chat::canChat( $wgUser ) ) {
-				ChatHelper::info( __METHOD__ . ': Method called - success' );
+				Chat::info( __METHOD__ . ': Method called - success' );
 				Wikia::setVar( 'OasisEntryControllerName', 'Chat' );
 				$wgOut->addModules( 'ext.Chat2' );
 				Chat::logChatWindowOpenedEvent();
 			} else {
-				ChatHelper::info( __METHOD__ . ': Method called - banned' );
+				Chat::info( __METHOD__ . ': Method called - banned' );
 				$wgOut->showErrorPage( 'chat-you-are-banned', 'chat-you-are-banned-text' );
 			}
 		} else {
 			// TODO: FIXME: Make a link on this page which lets the user login.
 			// TODO: FIXME: Make a link on this page which lets the user login.
 
-			ChatHelper::info( __METHOD__ . ': Method called - logged out' );
+			Chat::info( __METHOD__ . ': Method called - logged out' );
 			// $wgOut->permissionRequired( 'chat' ); // this is a really useless message, don't use it.
 			$wgOut->showErrorPage( 'chat-no-login', 'chat-no-login-text' );
